Safari-609.1.11.1/LayoutTests/accessibility/mac/aria-liveregions-notifications-expected.txt

changed text

new text element

This tests that ARIA live regions are sending out the correct notifications. We perform four operations (add, remove, change text, change alt tag), each one should trigger a live region notification

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS addedNotification1 is true
PASS addedNotification2 is true
PASS addedNotification3 is true
PASS addedNotification4 is true
PASS successfullyParsed is true

TEST COMPLETE
